requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-rest-swagger==2.2.0
django-simple-captcha==0.6.0
django-userforeignkey==0.5.0
djangorestframework==3.15.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
